adding sidebar and navbar folders & braking th erest to components
